+++ DesignToolbox/DesignToolbox/Pages/ThemeSelection/ThemeSelection.swift
@@ -13,6 +13,7 @@
 
 import OUDS
 import OUDSThemesOrange
+import OUDSThemesSosh
 import SwiftUI
 
 // MARK: - Extensions of OUDSTheme
@@ -34,6 +35,9 @@ extension OUDSTheme: @retroactive Identifiable, @retroactive Hashable {
     var name: String {
         if self is OrangeTheme {
             return "Orange"
+        }
+        if self is SoshTheme {
+            return "Sosh"
         }
         return String(describing: Self.self)
     }
@@ -67,8 +71,9 @@ final class ThemeProvider: ObservableObject {
 
     init() {
         let orangeTheme = OrangeTheme()
+        let soshTheme = SoshTheme()
         let defaultTheme = orangeTheme
-        themes = [orangeTheme]
+        themes = [orangeTheme, soshTheme]
 
         if let themeName = UserDefaults.standard.value(forKey: "themeName") as? String,
            let theme = themes.first(where: { $0.name == themeName })