+++ DesignToolbox/DesignToolbox/Pages/Tokens/Color/ColorTokenPage.swift
@@ -51,7 +51,9 @@ struct ColorTokenPage: View {
     private func illustrationForBackground() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Background.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -59,7 +61,9 @@ struct ColorTokenPage: View {
     private func illustrationForAction() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Action.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -67,7 +71,9 @@ struct ColorTokenPage: View {
     private func illustrationForAlways() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Always.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).isForbiddenValueColor() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -75,7 +81,9 @@ struct ColorTokenPage: View {
     private func illustrationForChart() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Chart.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -83,7 +91,9 @@ struct ColorTokenPage: View {
     private func illustrationForBorder() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Border.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -91,7 +101,9 @@ struct ColorTokenPage: View {
     private func illustrationForContent() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Content.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -99,7 +111,9 @@ struct ColorTokenPage: View {
     private func illustrationForDecorative() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Decorative.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).isForbiddenValueColor() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -107,7 +121,9 @@ struct ColorTokenPage: View {
     private func illustrationForOverlay() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Overlay.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -115,7 +131,9 @@ struct ColorTokenPage: View {
     private func illustrationForSurface() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Surface.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -123,7 +141,9 @@ struct ColorTokenPage: View {
     private func illustrationForOpacity() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Opacity.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).hasForbiddenColorValue() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -131,7 +151,9 @@ struct ColorTokenPage: View {
     private func illustrationForRepository() -> some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
             ForEach(NamedColor.Repository.allCases, id: \.rawValue) { namedColorToken in
-                Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                if !namedColorToken.token(from: theme).isForbiddenValueColor() {
+                    Illustration(token: namedColorToken.token(from: theme), name: namedColorToken.rawValue)
+                }
             }
         }
     }
@@ -159,7 +181,6 @@ struct ColorTokenPage: View {
 
         var body: some View {
             let colorRawToken = colorScheme == .dark ? lightValue : darkValue
-
             DesignToolboxTokenIllustration(tokenName: name, tokenValue: colorRawToken) {
                 Rectangle()
                     .fill(colorRawToken.color)