+++ CHANGELOG.md
@@ -14,6 +14,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Library] Update `controlItem` component to use `controlItemBorderRadius*` tokens (Orange-OpenSource/ouds-ios#777)
 - [Library] Update `switch` component to use `switchBorderRadius*` tokens (Orange-OpenSource/ouds-ios#780)
+- [Library] Sosh theme (Orange-OpenSource/ouds-ios#665)
+
+### Changed
+
+- [DesignToolbox] Available color modes and surfaces by theme for `colored surface` component
+- [Library] Color mode semantic tokens (tokens library v1.1.0)
+- [Library] Color multiple semantic tokens (tokens library v1.1.0)
 
 ## [0.16.0](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/0.15.0...0.16.0) - 2025-07-07